remove PowerPC examples and add ci target for running ci.sh on all platforms
This also reduces the set of examples to one per platform.  SWT 4.3
seems to have broken the paint example (or the paint example was
always broken, and SWT 4.3 just exposed the brokenness), so I dropped
that.  I also dropped the graphics example to reduce the smoke testing
burden.  The control example seems sufficient to demonstrate what
needs demonstrating.